refactor(audit-server): hoist db fixture
Gonna use it all over the place so put it in root conftest
